added the orders over time and the last days to the overview screen
