+++ app/(dashboard)/overview/last-sales.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function LastSales() {
+  return (
+    <div className="w-[30%] h-[700px] ml-8 bg-white rounded-md p-6 ">
+      OrdersOverTime
+    </div>
+  );
+}
+
+export default LastSales;
