+++ app/(dashboard)/overview/orders-overtime.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+function OrdersOverTime() {
+  return (
+    <div className="w-[70%] h-[700px] bg-white rounded-md p-6 ">OrdersOverTime</div>
+  );
+}
+
+export default OrdersOverTime;
